+++ src/common/stamp.rs
@@ -9,12 +9,12 @@ pub fn shuffle(seed: u128) -> String {
     let mash_up = time.as_millis() ^ time.as_nanos() ^ time.as_micros() ^ seed;
     let folded = mash_up.to_be_bytes().iter()
         .fold(0, |a, b| a ^ b);
-    return format!("{:02x?}", folded);
+    format!("{:02x?}", folded)
 }
 
 /// Returns a pseudorandom 8 character hexadecimal string.
 pub fn stamp(seed: u128) -> String {
-    return seed.to_string();
+    seed.to_string()
 
     // let mut combined = "".to_string();
     // for i in 0..4 {
